***Version 1.1****
This version now PCLake+, which extends the PCLake model to cover a wide range of freshwater lakes that differ in stratification
regime and climate-related processes. To this end, the model was extended with a hypolimnion layer that can be
invoked and configured by forcing functions or by simple built-in empirical relationships that impose stratification.
Further adjustments to the original PCLake model have been made with respect to the calculation of 1)
light irradiation in the water column, 2) evaporation processes and 3) phenology of macrophytes.
See the paper associated with the model for more information: https://www.sciencedirect.com/science/article/pii/S0304380019300201


Signed-off-by: Teurlincx <SvenT@nioo.knaw.nl>