Added PCLakeS+ to PCLake+
---Version 1.5.0---

-------------------

***1.5.0***
1) Added the PCLake(S)+ manual.
2) Added the PCLakeS+ excel file.
3) Merged the PCLakeS+ function.R with the functions.R. Also updated several functions for PCLake+ and improved function descriptions. The main differences are:
3.1) matchSwitchToForcing function - Added ReadQOutAbst switch. NB this switch is not part of the official PCLake+ code; Fixed bug in the ReadNutFrac switch, mNLoadPhyt was included in here, while it should not have been. NB this adjustment has not found its way into the PCLake+ file yet.
3.2) GetForcingAndInterpolate function - upgrade of this function with the capability to read from an excel or txt file.
3.3) Changed the name of adjustSedimentParamSettings function into adjustSedimentParamSettings_inclBank, as this reflects its function better. Also included an discription of the parameters, and added the options "clay_peat" and "clay_sand".
3.4) Added adjustSedimentParamSettings_PCLakeSplus, which allows a column name and does not include the bank (parameter fDOrgSoil).
3.5) Changed the function name of setParameters into setValues, as this reflects its function better.
3.6) Added the functions makeTranUpstTotFormulas and makeAuxNames_PCLakeSplus, both needed to run PCLakeS+.
3.7) Adjusted the RunModel function with additional solver settings (rtol and atol) and a default hmax setting of 0.5. Also simplified the return function and made the dyn.load more robust by using file.path rather than paste.
4) Merged the PCLakeS+ function_PCLake.R with the functions_PCLake.R. Also updated several functions for PCLake+ and improved function descriptions. The main differences are:
4.1) Added the prepInitials function, which can select the values needed for a restart.
4.2) Made the functions more robust by using file.path rather than paste to create file names.
4.3) Made object and argument names more intuitive.
4.4) In the PCModelReadDATMFile_PCLakePlus function, removed the dependency on java for reading excel files by switching to the readxl package. Added the option to read from either the excel file or a txt file.
4.5) PCModelReadDATMFile_PCLakeSPlus, a new function to read the PCLakeS+ sheets in excel.
4.6) Added PCModelAdjustlDATMandCPPfiles_PCLakeSplus to adjust settings and c++ files for PCLakeS+.
4.7) Added PCModelInitializeModel_PCLakeSplus, for PCLakeS+ model initialization.
4.8) Added PCModelNetworkRun, to run a PCLakeS+ model.
5) Added PCLakeS+ example work_case

***1.4.3***
Fixed a small error in the parameter sheet, and updated the default lake parameter set

***1.4.2***
Cleanup of existing files (no changes to model code)

*** version 1.4.1 ***
The PCLake+ model has been overhauled significantly to include:
1) An update to the water balance that allows users to specify outflow from both epi and hyp. Furthermore seepage and infiltration have been split into two seperate fluxes. Also, a precipitation term was added.
2) Measured time series parameters (forcings) for all water transportable substances were added that can be added to both epi and hyp. Existing parameters were renamed accordingly.
3) An error term for the water balance was added.
4) Groundwater infiltration was adjusted to increase setlling velocity when active
5) Nomenclature of parameters has been improved
6) Numerous small bug fixes including an error in terms of mixing depth, an error in light extinction and an error with constant sediment depth were fixed.

Many thanks to Lilith Kramer and Annette Janssen for much of this work.

*** version 1.3.2 ***
Added an updated manual with thanks to Annette Janssen

*** version 1.3.1 ***
Added the files used in the paper of [Chang et al 2020, Water] to the DATM folder

*** version 1.3.0 ***
--BOYUANT AND N-FIXING BLUEGREEN PHYTOPLANKTON--
-This version of PCLake+ includes the expansion of blue-green algae traits described in Chang et al., 2020 (accepted: Water). In short Bluegreen algae can now be given N-fixing capacity, allowing them to take up N in the form of N2. Also, Bluegreen algae can be given the capacity to become boyuant and create a floating layer at the surface (sDBlueSurf).
For more information we refer to Chang et al., 2020 (Water, in press).
Note: By default these traits are turned off.

--ERROR CATCHING--
-All states are now equiped with statements to stop them from going negative due to integration routine errors. If states become negative they are instead put to a NearZero value prior to being used further on in the model calculations.
Note: No error messages are given when this happens.

--DEPTH UPDATES--
-Depth has been reverted to a single sDepthW state, allowing for easier linkage with other models
-Epi and hypolimnion minimal depth issues have been fixed (could be encountered when flexible water depths were used)
-Added a way to force the system to follow a measured water depth as a time series, adjusting the water inflow/outflow accordingly (mDepthW parameter).

--N LOAD--
-measured time series for N load in the form of Phytoplankton can now be assigned to the model as a parameter
-This update fixes an N mass balance error that could occur when specifying the different fractions of nitrogen in the incoming load directly.

--NON SOURCE CODE UPDATES--
-Cleaned up descriptions
-Updated short names for DELWAQ/DUFLOW
-Small fix to the DATM exporter not exporting the last formula column in the derivative sheet

*** version 1.2.2 ***
Fixed issues with DATM import

*** version 1.2.1 ***
Fixed issues with bifurcations in PCLake+

***Version 1.2.0***
This new version of PCModel has been adjusted to allow all PCModel variants (PCLake, PCDitch, PCLake+) to be compiled seperately so that users will not need to recompile manually when they switch between PCModels.
PCLake+ has had fixes to the transport from Marsh zone to Lake zone (and vice versa), fixing an existing error.
PCLake+ has now been given a path variable in the Control Sheet pointing to the location of the gcc compiler.
PCDitch+ has now been given a path variable in the Control Sheet pointing to the location of the gcc compiler.
PCLake+ is now equipped with a DATM import/export, allowing the user to quickly exchange model configurations and results. Furthermore, this allows for users to include model setup easily and in a transferable manner with their papers and reports.
PCLake+ has had some minor fixes on fish stoichiometry that could lead to mass balance errors.
PCLake+ has been extended with the potential for dynamic plant height, by default this is turned off and users are asked to be mindful that it the model is not calibrated to use it.

***Version 1.1***
This version now PCLake+, which extends the PCLake model to cover a wide range of freshwater lakes that differ in stratification
regime and climate-related processes. To this end, the model was extended with a hypolimnion layer that can be
invoked and configured by forcing functions or by simple built-in empirical relationships that impose stratification.
Further adjustments to the original PCLake model have been made with respect to the calculation of 1)
light irradiation in the water column, 2) evaporation processes and 3) phenology of macrophytes.
See the paper associated with the model for more information: https://www.sciencedirect.com/science/article/pii/S0304380019300201

***Version 1.0***
PCLake, PCDitch DATM versions including user manuals and executables as previously available from the NIOO ftp download location.

Change log:
-PCDitch was rebuild on the PCLake DATM to fix issues with bifurcation analysis

Signed-off-by: pcmodel <pcmodel@nioo.knaw.nl>